pyproject.toml
[project]
name = "apitube"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
